pyproject.toml
[tool.poetry]
name = "ginear"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pyfzf = "^0.3.1"
requests = "^2.31.0"
mypy = "^1.6.0"
types-requests = "^2.31.0.8"

[tool.poetry.scripts]
gin = "ginear.ginear:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
